local-install/services/core/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pydantic==2.5.3
httpx==0.26.0